# backend/cmd/order-service/docs/swagger.yaml
basePath: /
definitions:
  domain.Order:
    properties:
      created_at:
        type: string
      currency:
        type: string
      id:
        type: string
      idempotency_key:
        type: string
      items:
        items:
          $ref: '#/definitions/domain.OrderItem'
        type: array
      shipping_address:
        $ref: '#/definitions/domain.ShippingAddress'
      status:
        $ref: '#/definitions/domain.OrderStatus'
      total_amount:
        description: cents
        type: integer
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  domain.OrderItem:
    properties:
      id:
        type: string
      order_id:
        type: string
      price:
        description: unit price in cents
        type: integer
      product_id:
        type: string
      quantity:
        type: integer
    type: object
  domain.OrderStatus:
    enum:
    - pending
    - confirmed
    - paid
    - cancelled
    - refunded
    type: string
    x-enum-varnames:
    - StatusPending
    - StatusConfirmed
    - StatusPaid
    - StatusCancelled
    - StatusRefunded
...
  domain.Product:
    properties:
      category:
        type: string
      created_at:
        type: string
      currency:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      price:
        description: unit price in cents
        type: integer
      sku:
        type: string
      status:
        $ref: '#/definitions/domain.ProductStatus'
      stock:
        type: integer
      updated_at:
        type: string
    type: object
  domain.ProductStatus:
    enum:
    - active
    - inactive
    - out_of_stock
    type: string
    x-enum-varnames:
    - StatusActive
    - StatusInactive
    - StatusOutOfStock
  domain.ShippingAddress:
    properties:
      city:
        type: string
      country:
        description: ISO 3166-1 alpha-2, e.g. "US"
        type: string
      name:
        type: string
      postal_code:
        type: string
      state:
        type: string
      street:
        type: string
    type: object
  domain.Wallet:
    properties:
      balance:
        description: cents
        type: integer
      created_at:
        type: string
      currency:
        type: string
      id:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
...
  handler.CancelOrderResponse:
    properties:
      message:
        example: order cancelled
        type: string
      order_id:
        example: ord-abc-123
        type: string
    type: object
...
  handler.CreateOrderBody:
    properties:
      currency:
        example: USD
        type: string
      items:
        items:
          $ref: '#/definitions/handler.CreateOrderItemBody'
        minItems: 1
        type: array
      shipping_address:
        $ref: '#/definitions/handler.ShippingAddressBody'
    required:
    - currency
    - items
    type: object
  handler.CreateOrderItemBody:
    properties:
      product_id:
        example: prod-abc-123
        type: string
      quantity:
        example: 2
        minimum: 1
        type: integer
    required:
    - product_id
    - quantity
    type: object
  handler.CreateProductRequest:
    properties:
      category:
        example: electronics
        type: string
      currency:
        example: USD
        type: string
      description:
        example: Noise-cancelling over-ear headphones
        type: string
      name:
        example: Wireless Headphones
        type: string
      price:
        description: cents
        example: 9999
        type: integer
      sku:
        example: HDPH-001
        type: string
      stock:
        example: 100
        minimum: 0
        type: integer
    required:
    - name
    - price
    - sku
    type: object
  handler.CreateWalletRequest:
    properties:
      currency:
        example: USD
        type: string
    required:
    - currency
    type: object
  handler.ErrorResponse:
    properties:
      error:
        example: something went wrong
        type: string
    type: object
  handler.GatewayChangePasswordBody:
    properties:
      new_password:
        example: newS3cret!
        type: string
      old_password:
        example: s3cret!Pass
        type: string
    type: object
...
  handler.GatewayErrorResponse:
    properties:
      error:
        example: service unavailable
        type: string
    type: object
...
  handler.GatewayLoginBody:
    properties:
      email:
        example: alice@example.com
        type: string
      password:
        example: s3cret!Pass
        type: string
    type: object
  handler.GatewayLogoutBody:
    properties:
      refresh_token:
        example: eyJhbGci...
        type: string
    type: object
...
  handler.GatewayRefreshBody:
    properties:
      refresh_token:
        example: eyJhbGci...
        type: string
    type: object
  handler.GatewayRegisterBody:
    properties:
      email:
        example: alice@example.com
        type: string
      name:
        example: Alice Smith
        type: string
      password:
        example: s3cret!Pass
        type: string
    type: object
  handler.GatewayTokenPairResponse:
    properties:
      access_token:
        example: eyJhbGci...
        type: string
      expires_in:
        example: 900
        type: integer
      refresh_token:
        example: eyJhbGci...
        type: string
    type: object
  handler.GatewayUserResponse:
    properties:
      created_at:
        example: "2026-03-13T00:00:00Z"
        type: string
      email:
        example: alice@example.com
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      name:
        example: Alice Smith
        type: string
    type: object
...
  handler.ListOrdersResponse:
    properties:
      limit:
        example: 20
        type: integer
      offset:
        example: 0
        type: integer
      orders:
        items:
          $ref: '#/definitions/domain.Order'
        type: array
    type: object
  handler.ListProductsResponse:
    properties:
      limit:
        example: 20
        type: integer
      offset:
        example: 0
        type: integer
      products:
        items:
          $ref: '#/definitions/domain.Product'
        type: array
      total:
        example: 42
        type: integer
    type: object
...
  handler.PaymentErrorResponse:
    properties:
      error:
        example: payment not found
        type: string
    type: object
  handler.PaymentResponse:
    properties:
      amount:
        example: 5000
        type: integer
      created_at:
        example: "2026-01-01T00:00:00Z"
        type: string
      currency:
        example: USD
        type: string
      failure_reason:
        example: card_declined
        type: string
      id:
        example: pay-uuid
        type: string
      method:
        example: card
        type: string
      order_id:
        example: order-uuid
        type: string
      status:
        allOf:
        - $ref: '#/definitions/domain.PaymentStatus'
        example: succeeded
      transaction_id:
        example: txn_abc123
        type: string
      updated_at:
        example: "2026-01-01T00:01:00Z"
        type: string
      user_id:
        example: user-uuid
        type: string
    type: object
  handler.ProductErrorResponse:
    properties:
      error:
        example: product not found
        type: string
    type: object
...
  handler.RefundResponse:
    properties:
      message:
        example: payment refunded
        type: string
      order_id:
        example: ord-uuid
        type: string
      payment_id:
        example: pay-uuid
        type: string
    type: object
...
  handler.ShippingAddressBody:
    properties:
      city:
        example: New York
        type: string
      country:
        example: US
        type: string
      name:
        example: Alice Smith
        type: string
      postal_code:
        example: "10001"
        type: string
      state:
        example: NY
        type: string
      street:
        example: 123 Main St
        type: string
    required:
    - city
    - country
    - name
    - postal_code
    - street
    type: object
...
  handler.UpdateProductRequest:
    properties:
      category:
        example: electronics
        type: string
      description:
        example: Updated description
        type: string
      name:
        example: Wireless Headphones Pro
        type: string
      price:
        example: 11999
        type: integer
      stock:
        example: 200
        type: integer
    type: object
...
  handler.WalletErrorResponse:
    properties:
      error:
        example: wallet not found
        type: string
    type: object
host: localhost:8081
info:
  contact: {}
  description: Manages order creation, retrieval, and status updates driven by payment
    events.
  title: Order Service
  version: "1.0"
paths:
  /api/v1/orders:
    get:
      description: Returns paginated orders for the authenticated user.
      parameters:
      - default: 20
        description: Max results (default 20, max 100)
        in: query
        name: limit
        type: integer
      - default: 0
        description: Pagination offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.ListOrdersResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List orders
      tags:
      - orders
    post:
      consumes:
      - application/json
      description: 'Create a new order. Idempotent: same Idempotency-Key returns the
        existing order.'
      parameters:
      - description: Client-generated unique key for idempotency
        in: header
        name: Idempotency-Key
        required: true
        type: string
      - description: Order details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.CreateOrderBody'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.Order'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create order
      tags:
      - orders
  /api/v1/orders/{id}:
    delete:
      description: Cancels a pending or confirmed order. Only the owner can cancel.
        Paid or refunded orders cannot be cancelled.
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.CancelOrderResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Cancel order
      tags:
      - orders
    get:
      description: Returns an order. Only the owner can access it.
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Order'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get order by ID
      tags:
      - orders
  /api/v1/payments/{id}:
    get:
      description: Returns payment details including status, amount, and transaction
        ID.
      parameters:
      - description: Payment ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.PaymentResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.PaymentErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.PaymentErrorResponse'
      security:
      - BearerAuth: []
      summary: Get payment
      tags:
      - payments
  /api/v1/payments/{id}/refund:
    post:
      description: Refunds a succeeded payment. Only the payment owner can request
        a refund. Publishes payments.refunded which credits the user's wallet.
      parameters:
      - description: Payment ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.RefundResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.PaymentErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.PaymentErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.PaymentErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.PaymentErrorResponse'
      security:
      - BearerAuth: []
      summary: Refund payment
      tags:
      - payments
  /api/v1/payments/order/{order_id}:
    get:
      description: Returns the payment associated with a specific order ID.
      parameters:
      - description: Order ID
        in: path
        name: order_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.PaymentResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.PaymentErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.PaymentErrorResponse'
      security:
      - BearerAuth: []
      summary: Get payment by order
      tags:
      - payments
  /api/v1/products:
    get:
      description: Returns paginated products. Filter by category, status, or a comma-separated
        list of IDs (used by order service for batch price lookup).
      parameters:
      - description: Comma-separated product IDs for batch lookup
        in: query
        name: ids
        type: string
      - description: Filter by category
        in: query
        name: category
        type: string
      - description: 'Filter by status: active, inactive, out_of_stock'
        in: query
        name: status
        type: string
      - default: 20
        description: Max results (default 20, max 100)
        in: query
        name: limit
        type: integer
      - default: 0
        description: Pagination offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.ListProductsResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ProductErrorResponse'
      summary: List products
      tags:
      - products
    post:
      consumes:
      - application/json
      description: Adds a new product to the catalog with an authoritative price and
        initial stock.
      parameters:
      - description: Product details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.CreateProductRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.Product'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ProductErrorResponse'
        "409":
          description: SKU already exists
          schema:
            $ref: '#/definitions/handler.ProductErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ProductErrorResponse'
      security:
      - BearerAuth: []
      summary: Create product
      tags:
      - products
  /api/v1/products/{id}:
    delete:
      description: Soft-deletes a product by setting its status to inactive. It remains
        in the DB.
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ProductErrorResponse'
      security:
      - BearerAuth: []
      summary: Deactivate product
      tags:
      - products
    get:
      description: Returns product details including current price and stock level.
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Product'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ProductErrorResponse'
      summary: Get product
      tags:
      - products
    patch:
      consumes:
      - application/json
      description: Partially updates a product. Only provided fields are changed.
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      - description: Fields to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateProductRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Product'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ProductErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ProductErrorResponse'
      security:
      - BearerAuth: []
      summary: Update product
      tags:
      - products
  /api/v1/wallet:
    get:
      description: Returns the authenticated user's wallet balance and details.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Wallet'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.WalletErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.WalletErrorResponse'
      security:
      - BearerAuth: []
      summary: Get wallet
      tags:
      - wallet
    post:
      consumes:
      - application/json
      description: Creates a new wallet for the authenticated user. Each user can
        have only one wallet.
      parameters:
      - description: Wallet details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.CreateWalletRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.Wallet'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.WalletErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.WalletErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.WalletErrorResponse'
      security:
      - BearerAuth: []
      summary: Create wallet
      tags:
      - wallet
  /auth/change-password:
    post:
      consumes:
      - application/json
      description: Changes the authenticated user's password and revokes all existing
        refresh tokens.
      parameters:
      - description: Old and new passwords
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.GatewayChangePasswordBody'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.GatewayErrorResponse'
      security:
      - BearerAuth: []
      summary: Change password
      tags:
      - auth
  /auth/login:
    post:
      consumes:
      - application/json
      description: Validates credentials and returns a token pair. Use the access_token
        as "Bearer <token>" in the Authorize dialog.
      parameters:
      - description: Credentials
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.GatewayLoginBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.GatewayTokenPairResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.GatewayErrorResponse'
      summary: Login
      tags:
      - auth
  /auth/logout:
    post:
      consumes:
      - application/json
      description: Revokes the given refresh token. Access tokens expire naturally
        after 15 min.
      parameters:
      - description: Refresh token to revoke
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.GatewayLogoutBody'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
      security:
      - BearerAuth: []
      summary: Logout
      tags:
      - auth
  /auth/me:
    get:
      description: Returns the authenticated user's profile.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.GatewayUserResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.GatewayErrorResponse'
      security:
      - BearerAuth: []
      summary: Get current user
      tags:
      - auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Exchanges a refresh token for a new token pair. The old refresh
        token is revoked immediately.
      parameters:
      - description: Refresh token
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.GatewayRefreshBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.GatewayTokenPairResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.GatewayErrorResponse'
      summary: Refresh tokens
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Creates a new user account and returns an access token (15 min)
        + refresh token (7 days).
      parameters:
      - description: Registration details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.GatewayRegisterBody'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handler.GatewayTokenPairResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.GatewayErrorResponse'
        "409":
          description: Email already taken
          schema:
            $ref: '#/definitions/handler.GatewayErrorResponse'
      summary: Register
      tags:
      - auth
...
securityDefinitions:
  BearerAuth:
    description: JWT token issued by the API Gateway
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"